+++ Source/WebCore/editing/Editor.cpp
@@ -36,6 +36,7 @@
 #include "CSSValueList.h"
 #include "CSSValuePool.h"
 #include "CachedResourceLoader.h"
+#include "CaretRectComputation.h"
 #include "ChangeListTypeCommand.h"
 #include "Chrome.h"
 #include "ChromeClient.h"
