+++ Source/WebCore/rendering/RenderBlockFlowInlines.h
@@ -19,6 +19,7 @@
 
 #pragma once
 
+#include "CaretRectComputation.h"
 #include "RenderBlockFlow.h"
 
 namespace WebCore {
