+++ Source/WebCore/rendering/RenderTableCell.h
@@ -271,15 +271,6 @@ inline unsigned RenderTableCell::rowIndex() const
     return row()->rowIndex();
 }
 
-inline bool RenderTableCell::isBaselineAligned() const
-{
-    if (auto alignContent = style().alignContent(); !alignContent.isNormal())
-        return alignContent.position() == ContentPosition::Baseline;
-
-    VerticalAlign va = style().verticalAlign();
-    return va == VerticalAlign::Baseline || va == VerticalAlign::TextBottom || va == VerticalAlign::TextTop || va == VerticalAlign::Super || va == VerticalAlign::Sub || va == VerticalAlign::Length;
-}
-
 inline RenderTableCell* RenderTableRow::firstCell() const
 {
     return downcast<RenderTableCell>(RenderBox::firstChild());
