+++ Source/WebCore/rendering/RenderTableCellInlines.h
@@ -21,6 +21,7 @@
 
 #include "RenderStyleInlines.h"
 #include "RenderTableCell.h"
+#include "StyleContentAlignmentData.h"
 
 namespace WebCore {
 
@@ -70,6 +71,15 @@ inline Length RenderTableCell::styleOrColLogicalWidth() const
     if (RenderTableCol* firstColumn = table()->colElement(col()))
         return logicalWidthFromColumns(firstColumn, styleWidth);
     return styleWidth;
+}
+
+inline bool RenderTableCell::isBaselineAligned() const
+{
+    if (auto alignContent = style().alignContent(); !alignContent.isNormal())
+        return alignContent.position() == ContentPosition::Baseline;
+
+    VerticalAlign va = style().verticalAlign();
+    return va == VerticalAlign::Baseline || va == VerticalAlign::TextBottom || va == VerticalAlign::TextTop || va == VerticalAlign::Super || va == VerticalAlign::Sub || va == VerticalAlign::Length;
 }
 
 } // namespace WebCore
